+++ modules/Student_Billing/StudentPayments.php
@@ -42,7 +42,7 @@ if ( ! empty( $_REQUEST['values'] )
 			);
 		}
 		elseif ( isset( $columns['AMOUNT'] )
-			&& $columns['AMOUNT'] != ''
+			&& is_numeric( $columns['AMOUNT'] )
 			&& $columns['PAYMENT_DATE'] )
 		{
 			$insert_columns = [
@@ -52,8 +52,6 @@ if ( ! empty( $_REQUEST['values'] )
 			];
 
 			$columns['FILE_ATTACHED'] = _savePaymentsFile( $id );
-
-			$columns['AMOUNT'] = preg_replace( '/[^0-9.-]/', '', $columns['AMOUNT'] );
 
 			// @since 11.2 Add CREATED_BY column to billing_fees & billing_payments tables
 			$columns['CREATED_BY'] = DBEscapeString( User( 'NAME' ) );